+++ DirNode.cpp
@@ -294,7 +294,7 @@ int DirNode::GET_LINK_NAME(Buffer* linkname, const char* path)
 		return -1;
 	}
 
-	linkname->_i = strlen(linkname->_v);
+	linkname->_i = (int) strlen(linkname->_v);
 	linkname->_l = linkname->_i;
 
 	return 0;
